pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "theus"
version = "3.0.0"
description = "Reliable, Rust-Powered Agentic AI Framework with Transactional Memory."
authors = [{name = "Do Huy Hoang", email = "dohuyhoangvn93@gmail.com"}]
readme = "README.md"
requires-python = ">=3.12"
keywords = ["ai", "agents", "rust", "memory", "transactional", "safety", "pop", "framework"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Rust",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typing-extensions>=4.0",
    "rich>=13.0",
    "questionary>=2.0",
]

[project.urls]
Homepage = "https://github.com/dohuyhoang93/theus"
Repository = "https://github.com/dohuyhoang93/theus"
Issues = "https://github.com/dohuyhoang93/theus/issues"

[project.scripts]
theus = "theus.cli:main"

[tool.maturin]
# "mixed" means we include both Python source code and Rust binary extension
features = ["pyo3/extension-module"]
# Python package lives in the theus/ subfolder
python-packages = ["theus"]
module-name = "theus_core"
